Better drive funtions
Changed the drive to move fword/back and strafing to move in cm. make shure to set the weele size in mm in robot.c and if it has a dicimal point divided accordingly as c++ dosent like decimal points. In addition to that more tightining up most notibly being autoselect display is a funtion so it isent a mess in main.cpp